+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/GoogleBigQueryInterpreter.scala
@@ -67,5 +67,15 @@ private[google2] class GoogleBigQueryInterpreter[F[_]: Sync: ContextShift: Timer
     )
   }
 
+  override def deleteDataset(datasetName: String): F[Boolean] = {
+    withLogging(
+      blockingF(Sync[F].delay[Boolean] {
+        client.delete(datasetName)
+      }),
+      None,
+      s"com.google.cloud.bigquery.BigQuery.delete(${datasetName})"
+    )
+  }
+
   private def blockingF[A](fa: F[A]): F[A] = blocker.blockOn(fa)
 }